move hash-to-file logic from TETagQuery.java to Utils.java
