+++ hashing/te-tag-query-java/com/facebook/threatexchange/TETagQuery.java
@@ -7,7 +7,6 @@ import org.json.simple.parser.JSONParser;
 import java.io.BufferedReader;
 import java.io.File;
 import java.io.FileNotFoundException;
-import java.io.FileOutputStream;
 import java.io.IOException;
 import java.io.InputStream;
 import java.io.InputStreamReader;
@@ -16,7 +15,6 @@ import java.lang.NumberFormatException;
 import java.net.URL;
 import java.util.ArrayList;
 import java.util.Arrays;
-import java.util.Base64;
 import java.util.List;
 import java.util.Map;
 import java.util.stream.Stream;
@@ -654,10 +652,10 @@ public class TETagQuery {
         System.out.println(hashFormatter.format(sharedHash, printHashString));
 
         if (hashDir != null) {
-          String path = hashDir + File.separator + sharedHash.hashID + hashTypeToFileSuffix(sharedHash.hashType);
+          String path = hashDir + File.separator + sharedHash.hashID + Utils.hashTypeToFileSuffix(sharedHash.hashType);
 
           try {
-            outputHashToFile(sharedHash, path, verbose);
+            Utils.outputHashToFile(sharedHash, path, verbose);
           } catch (FileNotFoundException e) {
             System.err.printf("FileNotFoundException: \"%s\".\n", path);
           } catch (IOException e) {
@@ -666,95 +664,6 @@ public class TETagQuery {
         }
       }
     }
-  }
-
-  public static String hashTypeToFileSuffix(String hashType) {
-    String suffix = ".hsh";
-    switch (hashType) {
-    case Constants.HASH_TYPE_PHOTODNA:
-      suffix = ".pdna";
-      break;
-    case Constants.HASH_TYPE_PDQ:
-      suffix = ".pdq";
-      break;
-    case Constants.HASH_TYPE_MD5:
-      suffix = ".md5";
-      break;
-    case Constants.HASH_TYPE_TMK:
-      suffix = ".tmk";
-      break;
-    }
-    return suffix;
-  }
-
-  public static void outputHashToFile(
-    SharedHash sharedHash,
-    String path,
-    boolean verbose)
-      throws FileNotFoundException, IOException
-  {
-    if (sharedHash.hashType.equals(Constants.HASH_TYPE_TMK)) {
-      outputTMKHashToFile(sharedHash, path, verbose);
-    } else {
-      outputNonTMKHashToFile(sharedHash, path, verbose);
-    }
-  }
-
-  // TMK hashes are binary data with data strucure in the TMK package's
-  // tmktypes.h. The string attributes are not 8-bit clean in the
-  // ThreatExchange API so the hashValue attribute is stored as base64-encoded,
-  // with one caveat: the first 12 bytes (which are ASCII) are stored as
-  // such, followed by a pipe delimiter, then the rest of the binary file
-  // base64-encdoed. Here we undo that encoding.
-  public static void outputTMKHashToFile(
-    SharedHash sharedHash,
-    String path,
-    boolean verbose)
-      throws FileNotFoundException, IOException
-  {
-    FileOutputStream fos = new FileOutputStream(path);
-    int hlen = sharedHash.hashValue.length();
-
-    String base64EncodedHash = sharedHash.hashValue;
-
-    byte[] bytes = base64EncodedHash.getBytes();
-    byte[] first = Arrays.copyOfRange(bytes, 0, 12);
-    byte[] rest = Base64.getDecoder().decode(Arrays.copyOfRange(bytes, 13, hlen));
-    fos.write(first);
-    fos.write(rest);
-    if (verbose) {
-      SimpleJSONWriter w = new SimpleJSONWriter();
-      w.add("path", path);
-      w.add("hash_type", sharedHash.hashType);
-      w.add("encoded_length", sharedHash.hashValue.length());
-      w.add("binary_length", first.length + rest.length);
-      System.out.println(w.format());
-      System.out.flush();
-    }
-    fos.close();
-  }
-
-  // Just write PhotoDNA, PDQ, MD5, etc. hash-data to the file contents.
-  public static void outputNonTMKHashToFile(
-    SharedHash sharedHash,
-    String path,
-    boolean verbose)
-      throws FileNotFoundException, IOException
-  {
-    FileOutputStream fos = new FileOutputStream(path);
-
-    byte[] bytes = sharedHash.hashValue.getBytes();
-    fos.write(bytes);
-    if (verbose) {
-      SimpleJSONWriter w = new SimpleJSONWriter();
-      w.add("path", path);
-      w.add("hash_type", sharedHash.hashType);
-      w.add("hash_length", sharedHash.hashValue.length());
-      System.out.println(w.format());
-      System.out.flush();
-    }
-
-    fos.close();
   }
 
   // ----------------------------------------------------------------
@@ -947,6 +856,7 @@ public class TETagQuery {
       o.printf("Required:\n");
       o.printf("-i|--indicator {...}   If indicator type is HASH_TMK this must be the\n");
       o.printf("                       path to a .tmk file, else the indicator text.\n");
+      o.printf("                       (This is under construction.)\n");
       o.printf("-t|--type {...}\n");
       o.printf("-d|--description {...}\n");
       o.printf("-l|--share-level {...}\n");
